:sparkles:(guestbook): add get_entry method and enhance guestbook routes
introduce get_entry method in guestbook_crud for fetching a specific guestbook entry
refactor guestbook routes to include public, authenticated, and admin routes
ensure proper authorization for deleting and flagging entries

+++ backend/src/wapi/guestbook/delete.rs
@@ -3,14 +3,25 @@ use axum::{
     response::IntoResponse,
 };
 
-use crate::{domain::models::Guest, errors::BResult, AppState};
+use crate::{
+    domain::models::Guest,
+    errors::{BResult, BackendError},
+    AppState,
+};
 
 pub async fn delete_entry(
     State(state): State<AppState>,
     guest: Guest,
     Path(id): Path<i64>,
 ) -> BResult<impl IntoResponse> {
-    // TODO: Check if the entry belongs to the guest or if the guest is an admin
-    state.guestbook_crud.delete_entry(id).await?;
-    Ok(())
+    let entry = state.guestbook_crud.get_entry(id).await?;
+
+    if entry.author_id == guest.id || guest.is_admin {
+        state.guestbook_crud.delete_entry(id).await?;
+        Ok(())
+    } else {
+        Err(BackendError::Forbidden(
+            "You are not allowed to delete this entry".to_string(),
+        ))
+    }
 }
